fix(day25): simple works as intended
But doesn't work.
The reason why it doesn't work is that it's too
greedy. jqt -> nvd -> cmg is a real issue.

To fix this, the algo needs to be adjusted with
more patience. It needs to resolve all options
before connecting.

The connection rule should be so that max conns
is first resolved, and any connection may be
created as long as it doesn't result in reduced
maxConns. Easier said than done.